Move ContentLayout into its own file
