+++ common/app/views/support/ContentLayout.scala
@@ -1,0 +1,41 @@
+package views.support
+
+import model._
+import org.jsoup.Jsoup
+
+object ContentLayout {
+  implicit class ContentLayout(content: model.Content) {
+    def showBottomSocialButtons: Boolean = {
+      content match {
+        case l: LiveBlog => true
+        case a: Article => Jsoup.parseBodyFragment(a.body).select("> *").text().length > 600
+        case i: ImageContent => false
+        case m: Media => false
+        case g: Gallery => false
+        case _ => true
+      }
+    }
+
+    def submetaBreakpoint: Option[String] = {
+      content match {
+        case a: LiveBlog => None
+        case a: Article if !a.hasSupporting => Some("leftcol")
+        case v: Video if v.standfirst.getOrElse("").length > 350 => Some("leftcol")
+        case a: Audio if a.body.getOrElse("").length > 800 => Some("leftcol")
+        case i: ImageContent if i.mainPicture.flatMap(_.largestEditorialCrop).exists(crop => crop.height / crop.width.toFloat > 0.5) => Some("wide")
+        case g: Gallery => Some("leftcol")
+        case _ => None
+      }
+    }
+
+    def tagTone: Option[String] = {
+      content match {
+        case l: LiveBlog => Some(l.visualTone)
+        case m: Media => Some("media")
+        case g: Gallery => Some("media")
+        case i: ImageContent if !i.isCartoon => Some("media")
+        case _ => None
+      }
+    }
+  }
+}
